bcsb-requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
alabaster==0.7.12
#asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.7
click==8.0.3
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==4.1b1
django-filter==22.1
djangorestframework==3.13.1
docutils==0.17.1
flake8==4.0.1
furl==2.1.3
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
imagesize==1.2.0
incremental==21.3.0
isort==5.10.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydash==5.1.0
pyflakes==2.4.0
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.17.2
pytest-cov==2.12.1
pytest-django==4.5.1
pytest-mock==3.6.1
python-keycloak==0.26.1
pytz==2021.3
regex==2021.11.2
requests==2.26.0
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-bluebrain-theme==0.2.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tomli==1.2.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zope.interface==5.4.0

setuptools~=57.0.0
